+++ src/pages/Todo/components/ImcompletedTodoDetailModal/NotificationDateTimeSetting/NotificationDateTimeSetting.module.css
@@ -7,29 +7,6 @@
   margin-right: 8px;
 }
 
-.timePicker {
-  display: flex;
-  gap: 8px;
-  border-radius: 12px;
-  padding: 8px;
-  background: var(--background-color);
-  color: var(--font-color);
-  overflow: hidden;
-}
-
-.timeSelect {
-  flex: 1;
-  height: 180px;
-  background: transparent;
-  color: inherit;
-  border: none;
-  overflow-y: auto;
-}
-
-.timeSelect option {
-  height: 36px;
-}
-
 :global(.react-datepicker),
 :global(.react-datepicker__header),
 :global(.react-datepicker-time__header),
