+++ src/pages/Todo/components/ImcompletedTodoDetailModal/NotificationDateTimeSetting/components/CustomTimeInput/CustomTimeInput.tsx
@@ -1,80 +1,5 @@
-import DatePicker from "react-datepicker";
-import "react-datepicker/dist/react-datepicker.css";
-import useNotificationDateTimeSettingViewModel from "./NotificationDateTimeSettingViewModel";
 import React from "react";
-import IconButton from "../../../../../../components/IconButton/IconButton";
-import { RxCross2 } from "react-icons/rx";
-import styles from "./NotificationDateTimeSetting.module.css";
-import { ja } from "date-fns/locale/ja";
-
-type Props = {
-  selectedDateTime: Date | null;
-  onChangeDateTime: (date: Date | null) => void;
-};
-
-const NotificationDateTimeSetting = ({
-  selectedDateTime,
-  onChangeDateTime,
-}: Props) => {
-  const { now, minTime, maxTime } =
-    useNotificationDateTimeSettingViewModel(selectedDateTime);
-
-  return (
-    <div className={styles.wrapper}>
-      <span className={styles.label}>リマインド:</span>
-      <DatePicker
-        selected={selectedDateTime}
-        onChange={onChangeDateTime}
-        customInput={<CustomDateDisplay />}
-        showTimeInput
-        customTimeInput={
-          <CustomTimeInput minTime={minTime} maxTime={maxTime} />
-        }
-        timeFormat="HH:mm"
-        dateFormat="M月d日 H時m分"
-        placeholderText="日時を選択"
-        minDate={now} // 過去の日付を選択不可にする
-        minTime={minTime}
-        maxTime={maxTime}
-        locale={ja}
-        withPortal
-        timeCaption="時間"
-      />
-      {selectedDateTime !== null && (
-        <IconButton onClick={() => onChangeDateTime(null)}>
-          <RxCross2 size={15} style={{ color: "var(--color-icon)" }} />
-        </IconButton>
-      )}
-    </div>
-  );
-};
-
-// カスタム表示用コンポーネントの型定義
-type CustomInputProps = {
-  value?: string;
-  onClick?: () => void;
-};
-
-// カスタムコンポーネント：<input> の代わりに <span> を使い、普通のテキスト表示にする
-const CustomDateDisplay = React.forwardRef<HTMLSpanElement, CustomInputProps>(
-  ({ value, onClick }, ref) => {
-    return (
-      <span
-        ref={ref}
-        onClick={onClick}
-        style={{
-          cursor: "pointer",
-          display: "inline-block",
-          userSelect: "none",
-          padding: "4px 8px",
-          borderBottom: "1px solid #ccc",
-        }}
-      >
-        {value || "日時を選択"}
-      </span>
-    );
-  }
-);
+import styles from "./CustomTimeInput.module.css";
 
 type CustomTimeInputProps = {
   date?: Date;
@@ -168,4 +93,4 @@ const CustomTimeInput = ({
   );
 };
 
-export default NotificationDateTimeSetting;
+export default CustomTimeInput;
